requirements.txt
streamlit>=1.30
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
requests>=2.31
python-dotenv>=1.0
pandas>=2.2
natasha>=1.6
pymupdf>=1.23.0
python-docx>=1.1.0
watchdog
easyocr
opencv-python-headless
numpy>=1.26
